requirements.tests.txt
faker
model-mommy
django-crypto-fields
edc-action-item
edc-appointment
edc-auth
edc-dashboard
edc-export
edc_identifier
edc-lab
edc-lab-dashboard
edc-locator
edc-metadata
edc-metadata-rules
edc-navbar
edc-notification
edc-offstudy
edc-pharmacy
edc-registration
edc-reference
edc-permissions
ambition-ae
ambition-lists
ambition-prn
ambition-rando
ambition-screening
ambition-subject
ambition-dashboard
